+++ src/herumi/index.ts
@@ -4,6 +4,9 @@ import {Signature} from "./signature.js";
 import {init, destroy} from "./context.js";
 import {IBls} from "../interface.js";
 import {functionalInterfaceFactory} from "../functional.js";
+
+await init();
+
 export * from "../constants.js";
 
 export {SecretKey, PublicKey, Signature, init, destroy};
@@ -15,8 +18,6 @@ export const bls: IBls = {
   Signature,
 
   ...functionalInterfaceFactory({SecretKey, PublicKey, Signature}),
-  init,
-  destroy,
 };
 
 export default bls;
