+++ src/index.ts
@@ -1,16 +1,14 @@
-import type {IBls, Implementation} from "./interface.js";
+import type {IBls} from "./interface.js";
 import {getImplementation} from "./getImplementation.js";
 
-export {IBls, Implementation, CoordType, PointFormat} from "./interface.js";
+// Thanks https://github.com/iliakan/detect-node/blob/master/index.esm.js
+const isNode = Object.prototype.toString.call(typeof process !== "undefined" ? process : 0) === "[object process]";
 
-// TODO: Use a Proxy for example to throw an error if it's not initialized yet
-export const bls: IBls = {} as IBls;
-export default bls;
-
-export async function init(impl: Implementation): Promise<void> {
-  // Using Object.assign instead of just bls = getImplementation()
-  // because otherwise the default import breaks. The reference is lost
-  // and the imported object is still undefined after calling init()
-  const blsImpl = await getImplementation(impl);
-  Object.assign(bls, blsImpl);
+let bls: IBls;
+try {
+  bls = await getImplementation(isNode ? "blst-native" : "herumi");
+} catch (e) {
+  bls = await getImplementation("herumi");
 }
+
+export default bls;
