+++ tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "outDir": "lib",
     "target": "es2019",
-    "module": "es2022",
+    "module": "esnext",
     "moduleResolution": "Node",
     "pretty": true,
     "lib": ["esnext.bigint", "DOM"],
